fix(web): resolve merge conflicts and add total_shares to RawPoolStatsResponse
