+++ packages/sdk/src/types/index.ts
@@ -42,6 +42,7 @@ export interface PoolStats {
   utilizationRateBps: number;
   totalYieldDistributed: bigint;
   activeInvoiceCount: number;
+  totalShares: bigint;
 }
 
 export interface LPPosition {
